Move mask matching logic to ProviderUtils and S3Helper
